refactor: move common components to components/common
