+++ pkgs/frontend/app/components/WorkspaceIcon.tsx
@@ -1,5 +1,5 @@
 import { FaPeopleGroup } from "react-icons/fa6";
-import { CommonIcon } from "./CommonIcon";
+import { CommonIcon } from "./common/CommonIcon";
 
 interface WorkspaceIconProps {
   workspaceImageUrl?: string;
