+++ pkgs/frontend/app/components/common/CommonInput.tsx
